V0.97 1.Add link management modal to subweb 2.Enhance multicata selection function in article modal 3.Enhance RTF editor 4.Some adjusts and bug fixes

+++ admin/func_link.php
@@ -7,6 +7,7 @@ if(empty($method)) $method = "list";
 $id = $req->getReq("id");
 $idx = $req->getReq("idx");
 $log_info = "";
+if(!$op_mode) $web_id = $setting['info']['web']['web_id'];
 
 switch($method) {
 	case "add":
@@ -15,13 +16,21 @@ switch($method) {
 		build_page($method);
 		break;
 	case "delete":
-		$log_info = $setting['language']['admin_func_link_delete'];
-		$db->Query("delete from ".$setting['db']['pre']."links where id = '$id'");
-		deleteCache("link");
+		if(!$op_mode && $web_id!=$_POST['web_id']) {
+			$goto_url = $setting['info']['self'];
+		} else {
+			$log_info = $setting['language']['admin_func_link_delete'];
+			$db->Query("delete from ".$setting['db']['pre']."links where id = '$id'");
+			deleteCache("link");
+		}
 		break;
 	case "add_ok":
 	case "edit_ok":
-		if(count($_POST) > 0) {
+		if(count($_POST) == 0) {
+			$goto_url = $setting['info']['self'];
+		} elseif(!$op_mode && $web_id!=$_POST['web_id']) {
+			$goto_url = $setting['info']['self'];
+		} else {
 			if($method=="add_ok") {
 				$log_info = $setting['language']['admin_func_link_add'];
 				$str_sql = $db->buildSQL($setting['db']['pre']."links", $_POST, "insert", "a");
@@ -31,8 +40,6 @@ switch($method) {
 			}
 			$db->Query($str_sql);
 			deleteCache("link");
-		} else {
-			$goto_url = $setting['info']['self'];
 		}
 		break;
 	default:
@@ -46,7 +53,7 @@ if(!empty($log_info)) {
 $mystep->pageEnd(false);
 
 function build_page($method) {
-	global $mystep, $req, $db, $tpl, $tpl_info, $setting, $id, $idx;
+	global $mystep, $req, $db, $tpl, $tpl_info, $setting, $id, $idx, $web_id;
 	
 	$tpl_info['idx'] = "func_link_".($method=="list"?"list":"input");
 	$tpl_tmp = $mystep->getInstance("MyTpl", $tpl_info);
@@ -55,16 +62,18 @@ function build_page($method) {
 		$order = $req->getGet("order");
 		$order_type = $req->getGet("order_type");
 		if(empty($order_type)) $order_type = "desc";
-		
 
-		$str_sql = "select count(*) as counter from ".$setting['db']['pre']."links";
-		if(!empty($idx)) $str_sql .= " where idx='".$idx."'";
+		$str_sql = "select count(*) as counter from ".$setting['db']['pre']."links where 1=1";
+		if(!empty($idx)) $str_sql .= " and idx='".$idx."'";
+		if(!empty($web_id)) $str_sql .= " and web_id='".$web_id."'";
 		$counter = $db->GetSingleResult($str_sql);
 		$page = $req->getGet("page");
 		list($page_arr, $page_start, $page_size) = GetPageList($counter, "?order={$order}&order_type={$order_type}", $page);
 		$tpl_tmp->Set_Variables($page_arr);
 
-		$str_sql = "select * from ".$setting['db']['pre']."links";
+		$str_sql = "select * from ".$setting['db']['pre']."links where 1=1";
+		if(!empty($idx)) $str_sql .= " and idx='".$idx."'";
+		if(!empty($web_id)) $str_sql .= " and web_id='".$web_id."'";
 		if(!empty($idx)) $str_sql .= " where idx='".$idx."'";
 		if(empty($order)) $order="id";
 		$str_sql.= " order by $order {$order_type}".(($order=="id")?"":", id desc");
@@ -89,6 +98,7 @@ function build_page($method) {
 		}
 		$tpl_tmp->Set_Variable('title', $setting['language']['admin_func_link_title']);
 		$tpl_tmp->Set_Variable('idx', $idx);
+		$tpl_tmp->Set_Variable('web_id', $web_id);
 	} else {
 		if($method == "edit") {
 			$db->Query("select * from ".$setting['db']['pre']."links where id='{$id}'");
@@ -99,10 +109,12 @@ function build_page($method) {
 				$mystep->show($tpl);
 				$mystep->pageEnd(false);
 			}
-			HtmlTrans(&$record);
+			$web_id = $record['web_id'];
 			$idx = $record['idx'];
+			HtmlTrans(&$record);
 		} else {
 			$record['id'] = "0";
+			$record['web_id'] = $web_id;
 			$record['idx'] = "";
 			$record['link_name'] = "";
 			$record['link_url'] = "http://";
@@ -114,6 +126,13 @@ function build_page($method) {
 		$tpl_tmp->Set_Variable('title', ($method == "add"?$setting['language']['admin_func_link_add']:$setting['language']['admin_func_link_edit']));
 		$tpl_tmp->Set_Variable('method', $method);
 		$tpl_tmp->Set_Variable('back_url', $req->getServer("HTTP_REFERER"));
+	}
+	$db->Free();
+	
+	$max_count = count($GLOBALS['website']);
+	for($i=0; $i<$max_count; $i++) {
+		$GLOBALS['website'][$i]['selected'] = $GLOBALS['website'][$i]['web_id']==$web_id?"selected":"";
+		$tpl_tmp->Set_Loop("website", $GLOBALS['website'][$i]);
 	}
 	$db->Free();
